(svn r13968) -Fix (r13923): 'cast from pointer to integer of different size' warning

+++ src/depot_gui.cpp
@@ -888,7 +888,7 @@ struct DepotWindow : Window {
 			/* Show tooltip window */
 			uint64 args[2];
 			args[0] = (whole_chain ? num : v->engine_type);
-			args[1] = (uint64)details;
+			args[1] = (uint64)(size_t)details;
 			GuiShowTooltips(whole_chain ? STR_DEPOT_VEHICLE_TOOLTIP_CHAIN : STR_DEPOT_VEHICLE_TOOLTIP, 2, args);
 		} else {
 			/* Show tooltip help */
